+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ All notable changes to Cellarium CAS client will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+dev - 2024-06-29
+----------------
+
+- Renamed data_preparation to preprocessing
+
+
 1.4.5 - 2024-06-03
 ------------------